pyproject.toml
  [build-system]
  requires = ["setuptools>=61.0", "wheel"]
  build-backend = "setuptools.build_meta"

  [project]
  name = "fact_checker"
  version = "0.1.0"
  description = "Мій дипломний проект для перевірки фактів"
  authors = [
    { name="Міщенко Данило", email="vantachput@gmail.com" },
  ]
  requires-python = ">=3.12"
  dependencies = [
      "aiofiles==25.1.0",
      "aiohappyeyeballs==2.6.1",
      "aiohttp==3.13.3",
      "aiosignal==1.4.0",
      "aiosqlite==0.22.1",
      "annotated-types==0.7.0",
      "anyio==4.12.1",
      "attrs==25.4.0",
      "black==25.12.0",
      "certifi==2026.1.4",
      "charset-normalizer==3.4.4",
      "click==8.3.1",
      "colorama==0.4.6",
      "distro==1.9.0",
      "dotenv==0.9.9",
      "eval_type_backport==0.2.2",
      "filelock==3.20.3",
      "frozenlist==1.8.0",
      "h11==0.16.0",
      "httpcore==1.0.9",
      "httpx==0.28.1",
      "idna==3.11",
      "jiter==0.12.0",
      "markdown-it-py==4.0.0",
      "mdurl==0.1.2",
      "multidict==6.7.0",
      "mypy_extensions==1.1.0",
      "numpy==2.4.1",
      "openai==2.15.0",
      "packaging==26.0",
      "pathspec==1.0.3",
      "pillow==11.3.0",
      "platformdirs==4.5.1",
      "propcache==0.4.1",
      "pydantic==2.12.5",
      "pydantic_core==2.41.5",
      "Pygments==2.19.2",
      "pyTelegramBotAPI==4.30.0",
      "python-dotenv==1.2.1",
      "python-telegram-bot==22.5",
      "pytokens==0.4.0",
      "pytz==2025.2",
      "requests==2.32.5",
      "rich==14.2.0",
      "shellingham==1.5.4",
      "sniffio==1.3.1",
      "tabulate==0.9.0",
      "together==1.5.35",
      "tqdm==4.67.1",
      "typer==0.19.2",
      "typing-inspection==0.4.2",
      "typing_extensions==4.15.0",
      "urllib3==2.6.3",
      "yarl==1.22.0"
  ]

  [tool.setuptools]
  # Автоматичний пошук пакетів у поточній папці
  packages = {find = {where = ["."]}}

[tool.ruff]
line-length = 88 

# Папки, які Ruff має ігнорувати
exclude = [
    ".git",
    ".venv",
    "venv",
    "env",
    "__pycache__",
    "build",
    "dist",
    "tests"
]
[tool.ruff.lint]
# Вмикаємо перевірку стилю (E), логічні помилки (F) та сортування імпортів (I)
select = [
    "E",    # pycodestyle (помилки стилю)
    "F",    # Pyflakes (логічні помилки, надійність)
    "I",    # isort (сортування імпортів)
    "PERF", # Perflint (продуктивність)
    "D",    # pydocstyle (документація)
]

[tool.ruff.lint.pydocstyle]
convention = "google"